src/main/java/ec/edu/ups/modelo/gestion_notas/Hora.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package ec.edu.ups.modelo.gestion_notas;

import java.io.Serializable;
import java.sql.Time;

/**
 *
 * @author claum
 */
public class Hora implements Serializable{
    private static final long serialVersionUID = 1L;
    private Time horaInicio;
    private Time horaFin;
    private String dia;

    public Hora() {
    }

    public Hora(Time horaInicio, Time horaFin, String dia) {
        this.horaInicio = horaInicio;
        this.horaFin = horaFin;
        this.dia = dia;
    }

    public Time getHoraInicio() {
        return horaInicio;
    }

    public void setHoraInicio(Time horaInicio) {
        this.horaInicio = horaInicio;
    }

    public Time getHoraFin() {
        return horaFin;
    }

    public void setHoraFin(Time horaFin) {
        this.horaFin = horaFin;
    }

    public String getDia() {
        return dia;
    }

    public void setDia(String dia) {
        this.dia = dia;
    }
    
    
}
